requirements.txt
pydantic
python-dotenv
langchain
langchain-core
langchain-groq